+++ libraries/lib-advanced-visualization/src/mCRendererContext.cpp
@@ -138,7 +138,7 @@ void CRendererContext::addChannel(const std::string& sChannelName, const float x
 	l_oChannelLocalisation.x = x * l_fInvNorm;
 	l_oChannelLocalisation.y = y * l_fInvNorm;
 	l_oChannelLocalisation.z = z * l_fInvNorm;
-	m_channelLookup.push_back(uint32_t(m_channelName.size()));
+	m_channelLookup.push_back(m_channelName.size());
 	m_channelName.push_back(sChannelName);
 	m_channelLocalisation.push_back(l_oChannelLocalisation);
 }
@@ -336,9 +336,9 @@ bool CRendererContext::getChannelLocalisation(const uint32_t index, float& x, fl
 	return true;
 }
 
-uint32_t CRendererContext::getChannelCount() const { return uint32_t(m_channelName.size()); }
+uint32_t CRendererContext::getChannelCount() const { return m_channelName.size(); }
 
-uint32_t CRendererContext::getSelectedCount() const { return uint32_t(m_channelLookup.size()); }
+uint32_t CRendererContext::getSelectedCount() const { return m_channelLookup.size(); }
 
 uint32_t CRendererContext::getSelected(const uint32_t index) const { return m_channelLookup[index]; }
