+++ plugins/examples/src/box-algorithms/ovpCBoxAlgorithmModifiableSettings.cpp
@@ -27,7 +27,7 @@ bool CBoxAlgorithmModifiableSettings::processClock(IMessageClock& /* messageCloc
 	//print settings values
 	for (size_t i = 0; i < m_SettingsValue.size(); ++i)
 	{
-		this->getLogManager() << LogLevel_Info << "Setting " << uint64_t(i) << " value is " << m_SettingsValue[i] << "\n";
+		this->getLogManager() << LogLevel_Info << "Setting " << i << " value is " << m_SettingsValue[i] << "\n";
 	}
 	this->getLogManager() << LogLevel_Info << "\n";
