+++ plugins/visualization/ovp-advanced-visualization/src/mTBoxAlgorithmInstantViz.hpp
@@ -176,7 +176,7 @@ namespace Mensia
 
 					if (nChannel == 0)
 					{
-						this->getLogManager() << LogLevel_Error << "Input stream " << uint32_t(i) << " has 0 channels\n";
+						this->getLogManager() << LogLevel_Error << "Input stream " << i << " has 0 channels\n";
 						return false;
 					}
